+++ Packages/com.unity.inputsystem/CHANGELOG.md
@@ -75,6 +75,7 @@ however, it has to be formatted properly to pass verification tests.
   * Rationale: These APIs are fairly low-level and not of general interest so having them out of `InputSystem` reduces the API surface visible to most users.
 - `InputDeviceChange.StateChanged` has been removed and is now a separate callback `InputState.onChange`.
   * Rationale: The other `InputDeviceChange` notifications are low-frequency whereas `StateChanged` is high-frequency. Putting them all on the same callback made adding a callback to `InputSystem.onDeviceChange` unnecessarily expensive.
+- `IInputStateCallbackReceiver` has been rewritten from scratch. Now has two simple methods `OnNextUpdate` and `OnEvent`. If implemented by a device, the device now has completely control over changing its own state. Use the `InputState.Change` methods to affect state changes while trigger state change monitors (e.g. for actions) correctly.
 
 ## [0.2.10-preview] - 2019-5-17